# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Room API",
        "description": "API endpoints for managing rooms.",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/api/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get all rooms",
                "operationId": "e33f8241fbefe09d5ae8a6d13db8cb6b",
                "responses": {
                    "200": {
                        "description": "List of rooms",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "operationId": "182a57048c0c30cc2a21978c5e4e5dab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created a new room",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a specific room",
                "operationId": "6917e3b2f885846bb93fdda516c776be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room details",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a room",
                "operationId": "4ac232229f4e67be48cebb13b46e4642",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Updated room",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room",
                "operationId": "0fcc6282f31b17139ccfa1b5995d009f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Room deleted"
                    }
                }
            }
        }
    }
}